Select a row only by clicking on checkbox; better styles for links.
This patch a workaround for “selecting a row only by clicking on checkbox”
problem (see https://github.com/callemall/material-ui/issues/3453).
And it's still possible to select a row if you click on empty space
in title column but it's much better than current behavior.

This patch also enhances styles for links in the table.